Sprint 3 Commit
Submit for Sprint 3. The app isn't working as designed and needs quite a bit more to make it feature functional.  It was a tough week with some family illness and a planned vacation and there just wan't time. I will continue to develop the code however and have it ready for the class presentation.